Return a list of tenants from ITenantsFactory
Originally, it returned a map with a key of data access. That is not sensible semantically
and it also caused problems in the ITenantsFactory.RemoveTenants method. The problem
was to deal with situation where multiple data access have same tenant and thus
we had to make sure not to remove tenant from pool twice.