Remove race condition
When multiple collection/class fixtures was instantiated at the same time,
it sometimes modifies the KeyValueStorage at the same time.

KeyValueStorage is a shared context and is a normal Dictionary, not a
Concurrent variant and would throw an exception during concurrent modification.

This change wraps each modification of KeyValueStorage to a critical section,
so only one thread can modify it at the same time. The keys the Sanctuary.xUnit
classes are modifying are distinct or values stored to a key are identical (in
case of multiple instantiation of a same collection/class fixture).