Add mapping of logical components to pooled ones
In the original code, components were mapped statically. There was no
way two tests could get different components. Instead, components
should be mapped dynamically.

The goal is to have a pool of components and to dynamically map
logical components to a component in the pool satisfies required
conditions for the component.

The dynamic mapping will allow us to utilize resources in more
distributed and efficient way.