Rename LogicalView to Template
This should help with understanding of new users what the purpose of
the config. The word "template" implies that it is used as a stamp to
create something repeatedly for each test, while "logical view" implies
that there already is something and we are just looking at it certain
way.