Add specification for a tenant in a template
Originally, there was only ability to set DataSource. That was very
limiting, because it didn't provide any opportunity to change tenant
options.

This change introduced ability to create a specification for a tenant,
where each type of a tenant can have different options (e.g. tier for
blobs and DOP for database).

The tenant specs are passed to a tenant factory durign tenant creation
and the factory creates a new tenant with desired characteristics.