Add specification for a component  in a template
Originally, components used static mapping. That is not a great
solution and thus we switched a dynamic mapping. The dynamic mapping
has a problem that we must ensure a correct component is selected. In
static mapping, that is done by developer. In dynamic mapping, it must
be done by library.

To make selection possible, each component has a set of conditions it
must satisfy. The conditions are passed to the pool and it finds a set
of components that satisfy the conditions.

Due to that, it's possible to utilize component pools even with larger/
dynamic number of components.