Add infra for parallel testing of ASP.NET Core
Originally, only tests that resolved and tested components from MSDI
in-process could be tested. That is because we used to use
Xunit.TestContext.Current as an ambient context.

The problem is that ASP.NET integration testing doesn't work with that
mode. The test server is running in a separate thread. The test creates
HttpClient and sends a request over loopback. Therefore, ambient
context is lost.

Instead, we are sending testId as a HTTP header with each request and a
middleware in test server receives it and sets the testId to the
ambinet context in test server request.